refactor: extract DbRegistry abstraction into @webiny/db
Move DbRegistry from local api-elasticsearch-tasks abstraction to a
proper DI abstraction + implementation in packages/db/src/feature/.
