+++ packages/db/package.json
@@ -17,6 +17,7 @@
   },
   "dependencies": {
     "@webiny/api": "0.0.0",
+    "@webiny/feature": "0.0.0",
     "type-fest": "^5.7.0"
   },
   "devDependencies": {
