+++ packages/db/src/feature/DbRegistry/index.ts
@@ -1,0 +1,2 @@
+export { DbRegistry } from "./abstractions.js";
+export { DbRegistryImpl } from "./DbRegistryImpl.js";
